[ProgressV2] fix lockable lessons expanding
