+++ apps/src/templates/sectionProgressV2/levelUrlUtils.jsx
@@ -1,0 +1,19 @@
+import queryString from 'query-string';
+
+export const navigateToLevelOverviewUrl = (levelUrl, studentId, sectionId) => {
+  if (!levelUrl) {
+    return null;
+  }
+  const params = {};
+
+  if (sectionId) {
+    params.section_id = sectionId;
+  }
+  if (studentId) {
+    params.user_id = studentId;
+  }
+  if (Object.keys(params).length) {
+    return `${levelUrl}?${queryString.stringify(params)}`;
+  }
+  return levelUrl;
+};
